Added dynamic iterations to problem spaces.
When experiments grow in size and complexity relative to a problem space
value, it can be difficult to define a constant number of iterations for
sampling that is neither too big nor too small for at least some of the
test cases.  This change allows for definition of an "override" on the
number of iterations when problem space values are also defined.