#64 Implemented TestFixture::Constants::NoProblemSpaceValue
This constant can be used to test if a problem space value was set or
not.

This is a patch for a null problem space value until a templated version
is implemented.  For now, all problem spaces are int64_t and can be used
as indexes into a vector<> which holds other types when necessary.