Move from CMake ModuleFind to CMake config-file
Updates `CMakeLists.txt`.

1. Updates installation of `include` folder. Now it is copying files in
project's `include` directory instead of directory itself. As stated by CMake
docs at
https://cmake.org/cmake/help/latest/command/install.html#directory "The last
component of each directory name is appended to the destination directory but
a trailing slash may be used to avoid this because it leaves the last component
empty". It clears the intention of operation as we need to copy to system
`include` directory. It is just a coincidance that our own interface directory
is called the same as install output one.

2. Updates installation of library. As we already knows that the target we are
installing is a library, we don't need to specify directories for other types
of targets.

3. Updates target importing file. The name is changed to `celero-targets.cmake`
to be guideline precise. The destination is changed to `lib/cmake/celero` as it
is directory where package config file will be placed. So it is convinient to
place files that will be accessed from config in the same directory.

4. Adds package config file creation and installation. This file has to be
passed to `configure_package_config_file()` function as it will unfold CMake specific macros.
After this it is copied to `lib/cmake/celero`. The full list of
possible output locations are listed at
https://cmake.org/cmake/help/latest/command/find_package.html#search-procedure.

Adds `config.cmake.in`.

1. This file will be configured by CMake, installed and searched for by
`find_package()` function.

Removes `FindCelero.cmake`.