Move dependency import to root CMakeLists.txt
Dependency import is better be done in root script file. The main idea behind
this is that you don't need to maintain documentation about required libraries.
Downstreams can open root CMakeLists.txt file and Ctrl+F all `find_package()`
or `find_dependency()`. The later one is more preferable as it output
meaningful message on any error. Some dependencies are required only on
certain occasion. For example, here we need GTest only if user checked
corresponding option. So for the first cmake configuration run it is possible
to check for GTest availability only after we introduce option with ON or OFF
set as default.