Fix cmake if command for compiler ids
CMake if command evaluates arguments. So it is recommended to pass variables
without `${}` syntax. The second part is better to be put in quotes. The
thorough discussion can be found at https://stackoverflow.com/questions/19982340/cmake-compare-to-empty-string-with-strequal-failed.

CMake sets `MSVC` to `true`. So there is no reason to compare it as string to
`CMAKE_CXX_COMPILER_ID`. CMake docs:
https://cmake.org/cmake/help/latest/variable/MSVC.html