CMake config file (#156)
* Move from CMake ModuleFind to CMake config-file

Updates `CMakeLists.txt`.

1. Updates installation of `include` folder. Now it is copying files in
project's `include` directory instead of directory itself. As stated by CMake
docs at
https://cmake.org/cmake/help/latest/command/install.html#directory "The last
component of each directory name is appended to the destination directory but
a trailing slash may be used to avoid this because it leaves the last component
empty". It clears the intention of operation as we need to copy to system
`include` directory. It is just a coincidance that our own interface directory
is called the same as install output one.

2. Updates installation of library. As we already knows that the target we are
installing is a library, we don't need to specify directories for other types
of targets.

3. Updates target importing file. The name is changed to `celero-targets.cmake`
to be guideline precise. The destination is changed to `lib/cmake/celero` as it
is directory where package config file will be placed. So it is convinient to
place files that will be accessed from config in the same directory.

4. Adds package config file creation and installation. This file has to be
passed to `configure_package_config_file()` function as it will unfold CMake specific macros.
After this it is copied to `lib/cmake/celero`. The full list of
possible output locations are listed at
https://cmake.org/cmake/help/latest/command/find_package.html#search-procedure.

Adds `config.cmake.in`.

1. This file will be configured by CMake, installed and searched for by
`find_package()` function.

Removes `FindCelero.cmake`.

* Use GNUInstallDirs CMake module

As stated at https://cmake.org/cmake/help/latest/command/install.html it is
recommended to use a path that begins with appropriate GNUInstallDirs variable.
This allows package maintainers to control the install destination by setting
the appropriate cache variables. The documentation on GNUInstallDirs can be
found at https://cmake.org/cmake/help/latest/module/GNUInstallDirs.html.

* Move dependency import to root CMakeLists.txt

Dependency import is better be done in root script file. The main idea behind
this is that you don't need to maintain documentation about required libraries.
Downstreams can open root CMakeLists.txt file and Ctrl+F all `find_package()`
or `find_dependency()`. The later one is more preferable as it output
meaningful message on any error. Some dependencies are required only on
certain occasion. For example, here we need GTest only if user checked
corresponding option. So for the first cmake configuration run it is possible
to check for GTest availability only after we introduce option with ON or OFF
set as default.

* Remove add_dependencies from experiments

There is no need to specify dependency order for libraries that are used in
`target_link_libraries()`. CMake handles this implicitly.

* Move PROJECT_NAME assignment to the beginning

The is no need to name project one way and later set it to another name. CMake
doesn't handle it differently between creation and assignment. This look like:

int x = 6;
// No usage of x.
x = 5;

The same goes with PROJECT_NAME:

project(CeleroProject)
// No usage of PROJECT_NAME.
set(PROJECT_NAME celero)

* Add warning suppression in experiments

Could not compile this projects with clang. Some variables are unused for the
testing purposes. So this commit make clang suppress warnings without affecting
the performance aspect of tests.

* Fix cmake if command for compiler ids

CMake if command evaluates arguments. So it is recommended to pass variables
without `${}` syntax. The second part is better to be put in quotes. The
thorough discussion can be found at https://stackoverflow.com/questions/19982340/cmake-compare-to-empty-string-with-strequal-failed.

CMake sets `MSVC` to `true`. So there is no reason to compare it as string to
`CMAKE_CXX_COMPILER_ID`. CMake docs:
https://cmake.org/cmake/help/latest/variable/MSVC.html